requirements.txt
allure-pytest
boto3~=1.16.43
botocore
botocore~=1.19.43
jsonschema~=3.2.0
mc-automation-tools>=1.0.46
mc-discrete-kit>=0.0.26
pyshp
pytest~=6.1.2
pytest-dependency
pytest-slack>=2.3.1
PyYAML~=5.4.1
requests~=2.25.0
setuptools~=50.3.2
shapely~=1.7.1
wheel
xmltodict~=0.12.0
